+++ packages/export/src/copy-webapp.js
@@ -2,27 +2,9 @@ const path = require('path');
 const glob = require('glob');
 
 const log = require('@home-gallery/logger')('export.copy.webapp');
+const { forEach } = require('@home-gallery/common');
 
 const copyFile = require('./copy-file');
-
-const copyFiles = (files, srcDir, dstDir, cb) => {
-  let i = 0;
-  const next = () => {
-    if (i === files.length) {
-      return cb();
-    }
-    const filename = files[i++];
-    copyFile(filename, srcDir, dstDir, (err) => {
-      if (err) {
-        const e = new Error(`Failed to copy ${filename} to ${dstDir}: ${err}`);
-        e.cause = err;
-        return cb(e);
-      }
-      next();
-    })
-  }
-  next();
-}
 
 const copyWebapp = (database, outputDirectory, basePath, cb) => {
   const t0 = Date.now();
@@ -33,12 +15,12 @@ const copyWebapp = (database, outputDirectory, basePath, cb) => {
     dot: true
   }, (err, files) => {
     if (err) {
-      log.error(`Could not copy webapp sources to ${dstDir}: ${err}`);
+      log.error(err, `Could not collect webapp sources of ${srcDir}: ${err}`);
       return cb(err);
     }
-    copyFiles(files, srcDir, dstDir, (err) => {
+    forEach(files, (filename, cb) => copyFile(filename, srcDir, dstDir, cb), (err) => {
       if (err) {
-        log.error(`Could not copy webapp sources to ${dstDir}: ${err}`)
+        log.error(err, `Could not copy webapp sources to ${dstDir}: ${err}`)
         return cb(err);
       }
       log.info(t0, `Copied webapp sources to ${dstDir}`);