+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- export: Fix webapp copy for export
 - index: Fix checksum option. Use `--no-checksum` to disable it
 
 ## [1.1.0] - 2021-06-22